feat(payment): PAYPAL-4867 refactoring, changes related to comments
