--- packages/core/src/bundles/checkout-headless-button.ts
@@ -1,3 +1,0 @@
-export { createTimeout } from '@bigcommerce/request-sender';
-
-export { createCheckoutHeadlessButtonInitializer } from '../checkout-buttons';
