+++ packages/core/src/bundles/headless-checkout-wallet.ts
@@ -1,0 +1,3 @@
+export { createTimeout } from '@bigcommerce/request-sender';
+
+export { createHeadlessCheckoutWalletInitializer } from '../checkout-buttons';
