+++ packages/core/src/checkout-buttons/checkout-button-initializer-options.ts
@@ -1,6 +1,4 @@
 export default interface CheckoutButtonInitializerOptions {
     host?: string;
     locale?: string;
-    storefrontJwtToken?: string;
-    siteLink?: string;
 }
