+++ packages/core/src/checkout-buttons/create-headless-checkout-wallet-initializer.ts
@@ -2,17 +2,18 @@ import { createRequestSender } from '@bigcommerce/request-sender';
 
 import { createCheckoutStore } from '../checkout';
 import { ConfigState } from '../config';
+import * as defaultCheckoutHeadlessWalletStrategyFactories from '../generated/checkout-headless-wallet-strategies';
 import { PaymentMethodActionCreator, PaymentMethodRequestSender } from '../payment';
 import { createPaymentIntegrationService } from '../payment-integration';
 
-import CheckoutButtonInitializerOptions from './checkout-button-initializer-options';
-import CheckoutHeadlessButtonInitializer from './checkout-headless-button-initializer';
-import CheckoutHeadlessButtonStrategyActionCreator from './checkout-headless-button-strategy-action-creator';
-import createCheckoutHeadlessButtonRegistryV2 from './create-checkout-headless-button-registry-v2';
+import createCheckoutButtonStrategyRegistryV2 from './create-checkout-button-registry-v2';
+import HeadlessCheckoutWalletInitializer from './headless-checkout-wallet-initializer';
+import HeadlessCheckoutWalletInitializerOptions from './headless-checkout-wallet-initializer-options';
+import HeadlessCheckoutWalletStrategyActionCreator from './headless-checkout-wallet-strategy-action-creator';
 
-export default function createCheckoutHeadlessButtonInitializer(
-    options?: CheckoutButtonInitializerOptions,
-): CheckoutHeadlessButtonInitializer {
+export default function createHeadlessCheckoutWalletInitializer(
+    options?: HeadlessCheckoutWalletInitializerOptions,
+): HeadlessCheckoutWalletInitializer {
     const { host, locale = 'en', storefrontJwtToken, siteLink } = options ?? {};
 
     const config: ConfigState = {
@@ -29,11 +30,14 @@ export default function createCheckoutHeadlessButtonInitializer(
     const store = createCheckoutStore({ config });
     const requestSender = createRequestSender({ host });
     const paymentIntegrationService = createPaymentIntegrationService(store);
-    const registryV2 = createCheckoutHeadlessButtonRegistryV2(paymentIntegrationService);
+    const registryV2 = createCheckoutButtonStrategyRegistryV2(
+        paymentIntegrationService,
+        defaultCheckoutHeadlessWalletStrategyFactories,
+    );
 
-    return new CheckoutHeadlessButtonInitializer(
+    return new HeadlessCheckoutWalletInitializer(
         store,
-        new CheckoutHeadlessButtonStrategyActionCreator(
+        new HeadlessCheckoutWalletStrategyActionCreator(
             registryV2,
             new PaymentMethodActionCreator(new PaymentMethodRequestSender(requestSender)),
         ),
