+++ packages/core/src/checkout-buttons/headless-checkout-wallet-initializer-options.ts
@@ -1,0 +1,6 @@
+export default interface HeadlessCheckoutWalletInitializerOptions {
+    host?: string;
+    locale?: string;
+    storefrontJwtToken?: string;
+    siteLink?: string;
+}
