--- packages/paypal-commerce-integration/src/paypal-commerce-credit/paypal-commerce-credit-headless-button-strategy.ts
@@ -1,228 +1,0 @@
-import {
-    CheckoutButtonInitializeOptions,
-    CheckoutButtonStrategy,
-    InvalidArgumentError,
-    MissingDataError,
-    MissingDataErrorType,
-    PaymentIntegrationService,
-} from '@bigcommerce/checkout-sdk/payment-integration-api';
-
-import PayPalCommerceIntegrationService from '../paypal-commerce-integration-service';
-import {
-    ApproveCallbackActions,
-    ApproveCallbackPayload,
-    PayPalCommerceButtonsOptions,
-    PayPalCommerceInitializationData,
-    ShippingAddressChangeCallbackPayload,
-    ShippingOptionChangeCallbackPayload,
-} from '../paypal-commerce-types';
-
-import PayPalCommerceCreditHeadlessButtonInitializeOptions, {
-    WithPayPalCommerceCreditHeadlessButtonInitializeOptions,
-} from './paypal-commerce-credit-headless-button-initialize-options';
-
-export default class PayPalCommerceCreditHeadlessButtonStrategy implements CheckoutButtonStrategy {
-    constructor(
-        private paymentIntegrationService: PaymentIntegrationService,
-        private paypalCommerceIntegrationService: PayPalCommerceIntegrationService,
-    ) {}
-
-    async initialize(
-        options: CheckoutButtonInitializeOptions &
-            WithPayPalCommerceCreditHeadlessButtonInitializeOptions,
-    ): Promise<void> {
-        const { paypalcommercecredit, containerId, methodId } = options;
-
-        if (!methodId) {
-            throw new InvalidArgumentError(
-                'Unable to initialize payment because "options.methodId" argument is not provided.',
-            );
-        }
-
-        if (!containerId) {
-            throw new InvalidArgumentError(
-                `Unable to initialize payment because "options.containerId" argument is not provided.`,
-            );
-        }
-
-        if (!paypalcommercecredit) {
-            throw new InvalidArgumentError(
-                `Unable to initialize payment because "options.paypalcommercecredit" argument is not provided.`,
-            );
-        }
-
-        const state = await this.paymentIntegrationService.loadCardEntity(
-            paypalcommercecredit.cartId,
-        );
-
-        const currencyCode = state.getCartOrThrow().currency.code;
-
-        await this.paypalCommerceIntegrationService.loadPayPalSdk(methodId, currencyCode, false);
-
-        this.renderButton(containerId, methodId, paypalcommercecredit);
-    }
-
-    deinitialize(): Promise<void> {
-        return Promise.resolve();
-    }
-
-    private renderButton(
-        containerId: string,
-        methodId: string,
-        paypalcommercecredit: PayPalCommerceCreditHeadlessButtonInitializeOptions,
-    ): void {
-        const { style, onEligibilityFailure } = paypalcommercecredit;
-
-        const paypalSdk = this.paypalCommerceIntegrationService.getPayPalSdkOrThrow();
-        const state = this.paymentIntegrationService.getState();
-        const paymentMethod =
-            state.getPaymentMethodOrThrow<PayPalCommerceInitializationData>(methodId);
-        const { isHostedCheckoutEnabled } = paymentMethod.initializationData || {};
-
-        const defaultCallbacks = {
-            createOrder: () =>
-                this.paypalCommerceIntegrationService.createPaymentOrderIntent(
-                    'paypalcommerce.paypalcredit',
-                ),
-            onApprove: ({ orderID }: ApproveCallbackPayload) =>
-                this.paypalCommerceIntegrationService.tokenizePayment(methodId, orderID),
-        };
-
-        const onShippingChangeCallbacks = {
-            onShippingAddressChange: (data: ShippingAddressChangeCallbackPayload) =>
-                this.onShippingAddressChange(data),
-            onShippingOptionsChange: (data: ShippingOptionChangeCallbackPayload) =>
-                this.onShippingOptionsChange(data),
-        };
-
-        const hostedCheckoutCallbacks = {
-            ...onShippingChangeCallbacks,
-            onApprove: (data: ApproveCallbackPayload, actions: ApproveCallbackActions) =>
-                this.onHostedCheckoutApprove(data, actions, methodId),
-        };
-
-        const fundingSources = [paypalSdk.FUNDING.PAYLATER, paypalSdk.FUNDING.CREDIT];
-        let hasRenderedSmartButton = false;
-
-        fundingSources.forEach((fundingSource) => {
-            if (!hasRenderedSmartButton) {
-                const buttonRenderOptions: PayPalCommerceButtonsOptions = {
-                    fundingSource,
-                    style: this.paypalCommerceIntegrationService.getValidButtonStyle(style),
-                    ...defaultCallbacks,
-                    ...(isHostedCheckoutEnabled && hostedCheckoutCallbacks),
-                };
-
-                const paypalButton = paypalSdk.Buttons(buttonRenderOptions);
-
-                if (paypalButton.isEligible()) {
-                    paypalButton.render(`#${containerId}`);
-                    hasRenderedSmartButton = true;
-                } else if (onEligibilityFailure && typeof onEligibilityFailure === 'function') {
-                    onEligibilityFailure();
-                }
-            }
-        });
-
-        if (!hasRenderedSmartButton) {
-            this.paypalCommerceIntegrationService.removeElement(containerId);
-        }
-    }
-
-    private async onHostedCheckoutApprove(
-        data: ApproveCallbackPayload,
-        actions: ApproveCallbackActions,
-        methodId: string,
-        onComplete?: () => void,
-    ): Promise<boolean> {
-        if (!data.orderID) {
-            throw new MissingDataError(MissingDataErrorType.MissingOrderId);
-        }
-
-        const state = this.paymentIntegrationService.getState();
-        const cart = state.getCartOrThrow();
-        const orderDetails = await actions.order.get();
-
-        try {
-            const billingAddress =
-                this.paypalCommerceIntegrationService.getBillingAddressFromOrderDetails(
-                    orderDetails,
-                );
-
-            await this.paymentIntegrationService.updateBillingAddress(billingAddress);
-
-            if (cart.lineItems.physicalItems.length > 0) {
-                const shippingAddress =
-                    this.paypalCommerceIntegrationService.getShippingAddressFromOrderDetails(
-                        orderDetails,
-                    );
-
-                await this.paymentIntegrationService.updateShippingAddress(shippingAddress);
-                await this.paypalCommerceIntegrationService.updateOrder();
-            }
-
-            await this.paymentIntegrationService.submitOrder({}, { params: { methodId } });
-            await this.paypalCommerceIntegrationService.submitPayment(methodId, data.orderID);
-
-            if (onComplete && typeof onComplete === 'function') {
-                onComplete();
-            }
-
-            return true; // FIXME: Do we really need to return true here?
-        } catch (error) {
-            if (typeof error === 'string') {
-                throw new Error(error);
-            }
-
-            throw error;
-        }
-    }
-
-    private async onShippingAddressChange(
-        data: ShippingAddressChangeCallbackPayload,
-    ): Promise<void> {
-        const address = this.paypalCommerceIntegrationService.getAddress({
-            city: data.shippingAddress.city,
-            countryCode: data.shippingAddress.countryCode,
-            postalCode: data.shippingAddress.postalCode,
-            stateOrProvinceCode: data.shippingAddress.state,
-        });
-
-        try {
-            // Info: we use the same address to fill billing and shipping addresses to have valid quota on BE for order updating process
-            // on this stage we don't have access to valid customer's address accept shipping data
-            await this.paymentIntegrationService.updateBillingAddress(address);
-            await this.paymentIntegrationService.updateShippingAddress(address);
-
-            const shippingOption = this.paypalCommerceIntegrationService.getShippingOptionOrThrow();
-
-            await this.paymentIntegrationService.selectShippingOption(shippingOption.id);
-            await this.paypalCommerceIntegrationService.updateOrder();
-        } catch (error) {
-            if (typeof error === 'string') {
-                throw new Error(error);
-            }
-
-            throw error;
-        }
-    }
-
-    private async onShippingOptionsChange(
-        data: ShippingOptionChangeCallbackPayload,
-    ): Promise<void> {
-        const shippingOption = this.paypalCommerceIntegrationService.getShippingOptionOrThrow(
-            data.selectedShippingOption.id,
-        );
-
-        try {
-            await this.paymentIntegrationService.selectShippingOption(shippingOption.id);
-            await this.paypalCommerceIntegrationService.updateOrder();
-        } catch (error) {
-            if (typeof error === 'string') {
-                throw new Error(error);
-            }
-
-            throw error;
-        }
-    }
-}
