--- packages/paypal-commerce-integration/src/paypal-commerce/paypal-commerce-headless-button-initialize-options.ts
@@ -1,7 +1,0 @@
-export default interface PayPalCommerceHeadlessButtonInitializeOptions {
-    cartId: string;
-}
-
-export interface WithPayPalCommerceHeadlessButtonInitializeOptions {
-    paypalcommerce?: PayPalCommerceHeadlessButtonInitializeOptions;
-}
