+++ packages/paypal-commerce-integration/src/paypal-commerce/paypal-commerce-headless-wallet-initialize-options.ts
@@ -1,0 +1,7 @@
+export default interface PaypalCommerceHeadlessWalletInitializeOptions {
+    cartId: string;
+}
+
+export interface WithPayPalCommerceHeadlessWalletInitializeOptions {
+    paypalcommerce?: PaypalCommerceHeadlessWalletInitializeOptions;
+}
